Add internal Chrome page edge case for SweetAlert popup
The SweetAlert warning popup fires in the options popup instead of the current tab when the current tab is an internal Chrome page. This is because script injections are not allowed on internal Chrome pages, so a fullscreen SweetAlert is not possible in such a case; firing the alert in the options popup (which is guaranteed to be visible to the user) is a valid workaround.