Add URL whitelist to redirect SweetAlerts on non-functioning websites
Certain websites (such as www.espn.com) display unpredictable behavior when it comes to firing the 'Clear All' SweetAlert, causing the alert to not display properly in certain cases. A whitelist circumvents this problem by redirecting the SweetAlert to within the options popup rather than the current webpage for certain whitelisted websites. This ensures that the SweetAlert will always be displayed to the user even for websites on which fullscreen SweetAlerts do not appear.