Make SweetAlerts fire in place by default
SweetAlerts fire in place by default instead of firing fullscreen (in the context of the current webpage). This implementation is far simpler and improves the UI since users no longer need to navigate away from the popup to interact with the SweetAlert popup. Furthermore, no more whitelisting is required--in-place SweetAlerts are guaranteed to fire since their context (the options popup) is guaranteed to be available on every webpage.