Add edge case handling for when selected text is whitespace
When the user selects whitespace and attempts to add it to Wordrive through the context menu, the resulting 'selectionText' property is undefined, which raises an error if string functions like trim() are called on it. By ensuring that the type of 'selectionText' isn't 'undefined' before proceeding to add to Wordrive, whitespace selections no longer get added to Wordrive and no longer have the potential to raise errors.