Add edge case handling for when edited words/URLs are duplicates
If users edit a word in such a way that the resulting word is a duplicate (i.e. the resulting word already exists in Wordrive), the two duplicate word entries and all their unique URLs are combined into a single word entry. If users edit a URL in a such a way that the resulting URL is a duplicate, one of the duplicate URLs is automatically removed.