Add URL verification for manual word adder
Users can only save and exit while using the manual word adder if the inputted URL is valid (empty URLs are considered valid, and whitespace is automatically trimmed on input). If an invalid URL is inputted, the 'Save' button will be blocked and Chrome will display an error tooltip when the user hovers over the URL input box. The URL input box will also be marked with a red border to indicate to the user that they must resolve the invalid URL before saving the entry. This URL verification feature reduces the possibility of having invalid URLs in Wordrive and encourages users to copy and paste URLs for reliability purposes (for example, to ensure that necessary protocols are included).