Repurpose URL drop-down menus into extended drop-down menus
URL drop-down menus are replaced with extended drop-down menus that encapsulate all functionality corresponding to a given Wordrive entry. Users may view and edit source URLs, reference URLs, and notes for each Wordrive entry. Furthermore, a timestamp is generated upon the addition of a new Wordrive entry and can be viewed within that entry's extended drop-down menu. The timestamp is generated with respect to the user's local machine time.