Add title fetching feature
URLs are labeled by the HTML titles of their corresponding webpages. If the HTML title cannot be fetched, the URL hostname is used as a fallback. In order to maximize UI feedback, URLs are labeled by default by their URL hostnames, which get replaced by fetched titles if available. If HTML titles are successfully fetched, they are saved and their corresponding URLs are marked as fetched to avoid future fetch requests (which may hamper performance).